pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hhru_parser"
version = "0.1.0"
description = "HTTP-парсер вакансий на hh.ru"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31,<3",
  "beautifulsoup4>=4.12,<5",
  "psycopg[binary]>=3.1,<4",
  "python-dotenv>=1.0,<2",
  "tqdm>=4.66,<5",
  "browser-cookie3>=0.19.1,<1",
]

[tool.setuptools.packages.find]
where = ["src"]
